refactor(app): refactor the app completely with module based approach
